fix: 新しいorganism COMP作成にともなう、既存のモーダルのmolecule COMPの修正
